backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
